+++ src/Enzo/hydro-mhd/riemann/CMakeLists.txt
@@ -11,7 +11,7 @@ add_library(riemann STATIC ${RIEMANN_SRC_FILES})
 #   src/Enzo/enzo.hpp which effectively includes all of the other header files
 #   in the Enzo/Cello layers. Since some of these may include headers from
 #   external libraries (e.g. src/Cello/_disk.hpp includes <hdf5.h>), we need to
-#   tell the CMake to link against these external libraries so that it can
+#   tell CMake to link against these external libraries so that it can
 #   inform the compiler to search for headers in appropriate directories for
 #   each dependency 
 target_link_libraries(riemann PRIVATE ${External_LIBS})
@@ -30,12 +30,13 @@ target_include_directories (riemann
                               # (also specifies the location of the library's
                               # private header files)
   PRIVATE
-  ${CMAKE_CURRENT_SOURCE_DIR}/.. # for including "enzo.hpp"
-  ${CMAKE_CURRENT_BINARY_DIR}/.. # for including "enzo.decl.hpp" (implicitly
-                                 # included by "enzo.hpp")
-  ${CMAKE_CURRENT_SOURCE_DIR}/../../Cello # for including "cello.hpp"
-  ${CMAKE_CURRENT_BINARY_DIR}/../../Cello # for including "simulation.decl.h"
-                                          # (implicitly included by "cello.hpp")
+  ${CMAKE_CURRENT_SOURCE_DIR}/../.. # for including "enzo.hpp"
+  ${CMAKE_CURRENT_BINARY_DIR}/../.. # for including "enzo.decl.hpp" (implicitly
+                                    # included by "enzo.hpp")
+  ${CMAKE_CURRENT_SOURCE_DIR}/../../../Cello # for including "cello.hpp"
+  ${CMAKE_CURRENT_BINARY_DIR}/../../../Cello # for including "simulation.decl.h"
+                                             # (implicitly included by
+                                             #  "cello.hpp")
   ${CHARM_INCLUDE_DIRS} # for including <charm++.h> (implicitly included by
                         # "cello.hpp")
 )
